fonts: Fix column width issue when writing in Japanese in Org tables
In Org tables, when I wrote Japanese text in one column, the table's alignment would be screwed
because the width of Japanese and English font characters is not proportional to the number of bytes
that they take. (For e.g. 1 English character takes up 1 byte, whereas 1 Japanese character takes up
2 bytes. But the width of all Japanese characters is *not* twice the width of all English
characters.

I added a recipe that I found on StackOverflow which explains the problem and provides a sample
solution. It defines the font for English and Japanese (/ Chinese) characters separately using
charsets. For each font that is defined by this recipe, it defines the `:size` of the font in pixels
as a floating point value from this table of font size pairs.

It uses specific fonts for English and the chinese characters. I believe that this is required
because the scaling table has been made with these fonts in mind. (i.e. 17:20 is a ratio that will
work only for this pair of fonts.)

It has some support for Windows and Linux, as it switches fonts based on the platform. However, it
does not seem to have any support for Mac OS.