x-list-fonts fails when called from terminal
The following comand fails.

$ emacs --batch -l ~/.emacs.d/init.el --eval '(x-list-fonts "Menlo 14")'
... init log ...
Window system is not in use or not initialized

This function works when the window-system is initialized but not when it is not initialized. I
could not find the alternative to this function. So, I replaced the order of the functions to get
batch mode working on Ubuntu. Batch mode will probably not work on work computers though.