sql-mode: Simplify SQL pretty printing using pg_format
Remove the imported configuration which used Perl's SQL formatter. It was strangea nd did not work
properly with common functions such as COUNT. It used to change COUNT(1) to COUNT( 1 ), which is not
recognized by MySQL 5.7. This gets read as a custom routine, which is undefined.

So, I switch to pg_format which provides better results.

https://github.com/darold/pgFormatter