org: link: Add function to copy the description and bind keys
There was already a function to remove the link from the link at point. But it was not bound to any
key and I had forgotten about its existence. Now, I have rebound the functions to the appropriate
keys, and I think it will work well now.

TODO(Refactor): I should be able to refactor the functions which remove the link, or replace the
link, or replace the description easily.