projectile: Enable caching of projectile cache
This will significantly improve the performance of projectile. But I am not sure when the cache will
be invalidated and recreated. So, I am only in the trial period for using this feature. If it
doesn't work out well (i.e. there is a file on disk that is not shown in projectile, then I will
revert this.)