shortcuts: Use a textfile based locate replacement
Locate is a good idea: It searches for all files inside the system. However, the current
implementation on Linux is strange and unhelpful. Its update takes too long and seems to run only
once a week which makes it useless.

A simple fzf-type interface to Locate would be very useful. The file index is stored inside the home
directory, which is often encrypted, thus preventing the common leakage pattern because the root
partition was unencrypted.

https://github.com/icyflame/locate-replacement