Remove Tree sitter and associated packages
I initially installed it with hope that it would work seamlessly. But the speed at which the library
parses even moderately sized files is too slow. So, it is not possible to use it smoothly on any
file size. Also, I did not use it as much because it is possible to go to all functions using a
simple Search string such /^func /. I might try this again later, if I have to start writing Golang
a lot more.